\name{runSWAT2009}
\alias{runSWAT2009}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
This runs the SWAT2009 executable in the current directory.
}
\description{
This function runs the SWAT2009 executable in the current directory.
}
\usage{
runSWAT2009(hist_wx,elev,rch)
}
\arguments{
  \item{hist_wx}{
Describe \code{hist_wx}
}
  \item{elev}{
Describe \code{elev}
}
  \item{rch}{
Describe \code{rch}
}
}
\author{
Daniel R. Fuka
}
