\name{SWATmodel-package}
\alias{SWATmodel-package}
\alias{SWATmodel}
\docType{package}
\title{
SWATmodel: A multi-OS / multi-platform implementation of the TAMU SWAT model in R.
}
\description{
The Soil and Water Assessment Tool is a river basin or
watershed scale model developed by Dr. Jeff Arnold for the USDA-ARS.
rev681}
\details{
\tabular{ll}{
Package: \tab SWATmodel\cr
Type: \tab Package\cr
Version: \tab 0.6.01\cr
Date: \tab 2011-12-22\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Included Functions:\cr
GSOD_history\cr

}
\author{
Fuka, DR and Easton, ZM\cr
Maintainer: Daniel Fuka <drfuka@vt.edu>
}
\keyword{ package }
