\name{runSWAT2009}
\alias{runSWAT2009}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
This runs the SWAT2009 executable in the current directory.
}
\description{
This function runs the SWAT2009 executable in the current directory.
}
\usage{
runSWAT2005()
}
\author{
Daniel R. Fuka
}
